stupid mistake:  putting 64 bit data into unsigned long

+++ c_src/eleveldb.cc
@@ -177,7 +177,7 @@ struct EleveldbOptions
     int m_LeveldbGroomingThreads;
 
     int m_TotalMemPercent;
-    int m_TotalMem;
+    size_t m_TotalMem;
 
     bool m_LimitedDeveloper;
     bool m_FadviseWillNeed;
@@ -199,7 +199,7 @@ struct EleveldbOptions
         syslog(LOG_ERR, "  m_LeveldbGroomingThreads: %d\n", m_LeveldbGroomingThreads);
 
         syslog(LOG_ERR, "         m_TotalMemPercent: %d\n", m_TotalMemPercent);
-        syslog(LOG_ERR, "                m_TotalMem: %d\n", m_TotalMem);
+        syslog(LOG_ERR, "                m_TotalMem: %zd\n", m_TotalMem);
 
         syslog(LOG_ERR, "        m_LimitedDeveloper: %s\n", (m_LimitedDeveloper ? "true" : "false"));
         syslog(LOG_ERR, "         m_FadviseWillNeed: %s\n", (m_FadviseWillNeed ? "true" : "false"));
@@ -237,12 +237,12 @@ ERL_NIF_TERM parse_init_option(ErlNifEnv* env, ERL_NIF_TERM item, EleveldbOption
     {
         if (option[0] == eleveldb::ATOM_TOTAL_LEVELDB_MEM)
         {
-            unsigned long memory_sz;
+            size_t memory_sz;
             if (enif_get_ulong(env, option[1], &memory_sz))
             {
                 if (memory_sz != 0)
                 {
-                     opts.m_TotalMem = memory_sz;
+                    opts.m_TotalMem = memory_sz;
                 }
             }
         }
